+++ Qt_OpenGL_5_2/game.h
@@ -17,7 +17,7 @@
 #include "computerplayer.h"
 #include "TexturesManager/texturesmanager.h"
 #include "EnvGenerator/EnvGenerator.h"
-#include "Physics/physics.h"
+#include "Engine/physics.h"
 #include "drawableobject.h"
 #include "camera.h"
 #include "light.h"
